Add ability to specify export prefix
This is required to build a shared library on Windows, where the linker will only export the symbol if it has the `__declspec(dllexport)` attribute.

Since using it as a shared library isn't the default usecase of the lib, I've left it blank by default, with the ability to override it through `#define SDL_GPU_SHADERCROSS_EXPORT __declspec(dllexport)` before importing the header's implementation section, or through the user's build system.